+++ bloop/__init__.py
@@ -57,4 +57,4 @@ __all__ = [
     # Misc
     "Condition", "QueryIterator", "ScanIterator", "Stream", "missing"
 ]
-__version__ = "2.0.1"
+__version__ = "2.1.0"